Rename file sttructure
